+++ branches/31104branch/src/daily/shopsnstuff.cpp
@@ -131,7 +131,7 @@ void hospital(int loc)
    locatesquad(activesquad,loc);
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)
       {
@@ -197,7 +197,7 @@ char gunselect(creaturest *cr,short &gun,bool legal=1)
    gun=-1;
 
    vector<enum Weapons> guntype;
-   for(unsigned int a=0;a<WEAPONNUM;a++)
+   for(int a=0;a<WEAPONNUM;a++)
    {
       switch(a)
       {
@@ -223,7 +223,7 @@ char gunselect(creaturest *cr,short &gun,bool legal=1)
       }
    }
 
-   unsigned page=0;
+   int page=0;
 
    char str[200];
 
@@ -241,9 +241,9 @@ char gunselect(creaturest *cr,short &gun,bool legal=1)
       addstr("----PRODUCT NAME-----------------------PRICE------------------------------------");
 
       int y=2;
-      for(unsigned p=page*19;p<guntype.size()&&p<page*19+19;p++)
+      for(int p=page*19;p<guntype.size()&&p<page*19+19;p++)
       {
-         unsigned int price = gun_price(guntype[p]);
+         int price = gun_price(guntype[p]);
          if(funds<price)set_color(COLOR_BLACK,COLOR_BLACK,1);
          else set_color(COLOR_WHITE,COLOR_BLACK,0);
          move(y,0);
@@ -281,10 +281,10 @@ char gunselect(creaturest *cr,short &gun,bool legal=1)
 
       if(c>='a'&&c<='s')
       {
-         unsigned p=page*19+(int)(c-'a');
+         int p=page*19+(int)(c-'a');
          if(p<guntype.size())
          {
-            unsigned price = gun_price(guntype[p]);
+            int price = gun_price(guntype[p]);
             if(price < funds)
             {
                gun=guntype[p];
@@ -311,7 +311,7 @@ void armsdealer(int loc)
    locatesquad(activesquad,loc);
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)
       {
@@ -424,7 +424,7 @@ void armsdealer(int loc)
             }
 
             //DROP ALL CLIPS THAT DON'T WORK
-            for(unsigned int cl=0;cl<CLIPNUM;cl++)
+            for(int cl=0;cl<CLIPNUM;cl++)
             {
                if(cl==ammotype(activesquad->squad[buyer]->weapon.type))continue;
 
@@ -589,7 +589,7 @@ void pawnshop(int loc)
    locatesquad(activesquad,loc);
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)
       {
@@ -927,7 +927,7 @@ void pawnshop(int loc)
          if((c=='w'||c=='c'||c=='l'||c=='a'||c=='f')&&
             activesquad->loot.size()>0)
          {
-            unsigned long fenceamount=0;
+            int fenceamount=0;
 
             if(c=='f')fenceamount=fenceselect();
             else
@@ -1170,7 +1170,7 @@ void pawnshop(int loc)
             }
 
             //DROP ALL CLIPS THAT DON'T WORK
-            for(unsigned int cl=0;cl<CLIPNUM;cl++)
+            for(int cl=0;cl<CLIPNUM;cl++)
             {
                if(cl==ammotype(activesquad->squad[buyer]->weapon.type))continue;
 
@@ -1321,7 +1321,7 @@ void pawnshop(int loc)
             }
 
             //DROP ALL CLIPS THAT DON'T WORK
-            for(unsigned int cl=0;cl<CLIPNUM;cl++)
+            for(int cl=0;cl<CLIPNUM;cl++)
             {
                if(cl==ammotype(activesquad->squad[buyer]->weapon.type))continue;
 
@@ -1374,7 +1374,7 @@ void deptstore(int loc)
    locatesquad(activesquad,loc);
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)
       {
@@ -1530,7 +1530,7 @@ void deptstore(int loc)
          }
 
          //DROP ALL CLIPS THAT DON'T WORK
-         for(unsigned int cl=0;cl<CLIPNUM;cl++)
+         for(int cl=0;cl<CLIPNUM;cl++)
          {
             if(cl==ammotype(activesquad->squad[buyer]->weapon.type))continue;
 
@@ -1572,7 +1572,7 @@ void halloweenstore(int loc)
    locatesquad(activesquad,loc);
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)
       {
@@ -1950,7 +1950,7 @@ void halloweenstore(int loc)
          }
 
          //DROP ALL CLIPS THAT DON'T WORK
-         for(unsigned int cl=0;cl<CLIPNUM;cl++)
+         for(int cl=0;cl<CLIPNUM;cl++)
          {
             if(cl==ammotype(activesquad->squad[buyer]->weapon.type))continue;
 
@@ -1989,7 +1989,7 @@ char maskselect(creaturest *cr,short &mask)
    mask=-1;
 
    vector<int> masktype;
-   for(unsigned int a=0;a<MASKNUM;a++)
+   for(int a=0;a<MASKNUM;a++)
    {
       switch(a)
       {
@@ -2014,7 +2014,7 @@ char maskselect(creaturest *cr,short &mask)
       }
    }
 
-   unsigned page=0;
+   int page=0;
 
    char str[200];
 
@@ -2031,8 +2031,8 @@ char maskselect(creaturest *cr,short &mask)
       move(1,0);
       addstr("----PRODUCT NAME-----------------------DESCRIPTION------------------------------");
 
-      unsigned int y=2;
-      for(unsigned int p=page*19;p<masktype.size()&&p<page*19+19;p++)
+      int y=2;
+      for(int p=page*19;p<masktype.size()&&p<page*19+19;p++)
       {
          set_color(COLOR_WHITE,COLOR_BLACK,0);
          move(y,0);
@@ -2070,7 +2070,7 @@ char maskselect(creaturest *cr,short &mask)
 
       if(c>='a'&&c<='s')
       {
-         unsigned p=page*19+(int)(c-'a');
+         int p=page*19+(int)(c-'a');
          if(p<masktype.size())
          {
             mask=masktype[p];
@@ -2092,17 +2092,17 @@ char maskselect(creaturest *cr,short &mask)
 
 
 /* pick stuff to fence */
-unsigned long fenceselect(void)
+int fenceselect(void)
 {
-   unsigned long ret=0;
+   int ret=0;
 
    consolidateloot(activesquad->loot);
 
-   unsigned page=0;
+   int page=0;
 
    vector<int> selected;
    selected.resize(activesquad->loot.size());
-   for(unsigned int s=0;s<selected.size();s++)selected[s]=0;
+   for(int s=0;s<selected.size();s++)selected[s]=0;
 
    do
    {
@@ -2126,7 +2126,7 @@ unsigned long fenceselect(void)
       int x=1,y=10;
       char str[200],str2[200];
 
-      for(unsigned int l=page*18;l<activesquad->loot.size()&&l<page*18+18;l++)
+      for(int l=page*18;l<activesquad->loot.size()&&l<page*18+18;l++)
       {
          if(selected[l])set_color(COLOR_GREEN,COLOR_BLACK,1);
          else set_color(COLOR_WHITE,COLOR_BLACK,0);
@@ -2212,7 +2212,7 @@ unsigned long fenceselect(void)
 
       if(c>='a'&&c<='r')
       {
-         unsigned int slot=c-'a'+page*18;
+         int slot=c-'a'+page*18;
 
          if(slot>=0&&slot<activesquad->loot.size())
          {
@@ -2306,9 +2306,9 @@ unsigned long fenceselect(void)
 
 
 /* value of stuff to fence */
-unsigned long fencevalue(itemst &it)
+int fencevalue(itemst &it)
 {
-   unsigned long fenceamount=0;
+   int fenceamount=0;
 
    switch(it.type)
    {
@@ -2436,7 +2436,7 @@ void choose_buyer(short &buyer)
    party_status=-1;
 
    int partysize=0;
-   for(unsigned int p=0;p<6;p++)
+   for(int p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)partysize++;
    }
